# api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/remittances": {
      "post": {
        "operationId": "RemittancesController_create",
        "summary": "Record an inward remittance",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRemittanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Remittance recorded",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "id": "665f2a119a4e1c0012ab55ef",
                      "orgId": "665f1a009a4e1c0012ab1100",
                      "invoiceId": "665f1b2c9a4e1c0012ab34cd",
                      "amountReceived": "5000.00",
                      "currency": "USD",
                      "receivedDate": "2026-06-10",
                      "purposeCode": "P0802",
                      "fircUrl": null,
                      "createdAt": "2026-06-10T09:30:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Remittances"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "RemittancesController_list",
        "summary": "List remittances for an invoice",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "query",
            "description": "Invoice id to list remittances for",
            "schema": {
              "example": "665f1b2c9a4e1c0012ab34cd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remittances",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": [
                      {
                        "id": "665f2a119a4e1c0012ab55ef",
                        "orgId": "665f1a009a4e1c0012ab1100",
                        "invoiceId": "665f1b2c9a4e1c0012ab34cd",
                        "amountReceived": "5000.00",
                        "currency": "USD",
                        "receivedDate": "2026-06-10",
                        "purposeCode": "P0802",
                        "fircUrl": null,
                        "createdAt": "2026-06-10T09:30:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Remittances"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/remittances/{id}/firc": {
      "post": {
        "operationId": "RemittancesController_uploadFirc",
        "summary": "Upload FIRC / e-FIRA",
        "description": "Attach the bank FIRC document (PDF or image, max 10 MB) to a remittance.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Remittance id",
            "schema": {
              "example": "665f2a119a4e1c0012ab55ef",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "firc"
                ],
                "properties": {
                  "firc": {
                    "type": "string",
                    "format": "binary",
                    "description": "FIRC file (pdf/png/jpg, ≤ 10 MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "FIRC attached",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "id": "665f2a119a4e1c0012ab55ef",
                      "orgId": "665f1a009a4e1c0012ab1100",
                      "invoiceId": "665f1b2c9a4e1c0012ab34cd",
                      "amountReceived": "5000.00",
                      "currency": "USD",
                      "receivedDate": "2026-06-10",
                      "purposeCode": "P0802",
                      "fircUrl": "/api/v1/remittances/665f2a119a4e1c0012ab55ef/firc",
                      "createdAt": "2026-06-10T09:30:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Remittance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Remittances"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "RemittancesController_downloadFirc",
        "summary": "Download FIRC",
        "description": "Streams the stored FIRC document inline (bypasses the JSON envelope).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Remittance id",
            "schema": {
              "example": "665f2a119a4e1c0012ab55ef",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary FIRC document stream",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "FIRC not found for this remittance",
            "content": {
              "application/pdf": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              },
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Remittances"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Crossbill API",
    "description": "Crossbill backend API — GST / FEMA compliant invoicing & remittance for Indian exporters.\n\n**Conventions**\n- Base path: `/api/v1`\n- Success envelope: `{ \"success\": true, \"data\": ... }` (list endpoints add a `meta` object).\n- Error envelope: `{ \"success\": false, \"error\": { \"code\", \"message\", \"correlationId\" }, \"timestamp\", \"path\" }`.\n- Auth: send `Authorization: Bearer <accessToken>` (obtained from `/auth/login` or `/auth/register`).\n- The refresh token is delivered as the `crossbill_rt` httpOnly cookie and is read automatically by `/auth/refresh`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Base URL (includes /api/v1 prefix)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the accessToken returned by /auth/login or /auth/register."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "crossbill_rt",
        "description": "httpOnly refresh-token cookie set by the auth endpoints; used by /auth/refresh."
      }
    },
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable error message"
          },
          "correlationId": {
            "type": "string",
            "example": "7f3c1e9a-2b4d-4c8e-9a1b-1f2e3d4c5b6a",
            "description": "Request correlation id"
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-06-12T10:15:30.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/clients"
          }
        },
        "required": [
          "success",
          "error",
          "timestamp",
          "path"
        ]
      },
...
      "CreateRemittanceDto": {
        "type": "object",
        "properties": {
          "invoiceId": {
            "type": "string",
            "description": "Id of the invoice this remittance settles",
            "example": "665f1b2c9a4e1c0012ab34cd"
          },
          "amountReceived": {
            "type": "number",
            "description": "Amount received in the remittance currency",
            "example": 5000,
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "description": "Currency of the received amount (ISO 4217)",
            "example": "USD"
          },
          "receivedDate": {
            "type": "string",
            "description": "Date the funds were received (ISO 8601)",
            "format": "date",
            "example": "2026-06-10"
          },
          "purposeCode": {
            "type": "string",
            "description": "RBI purpose code for the inward remittance",
            "enum": [
              "P0802",
              "P0801",
              "P0803",
              "P0806",
              "P0807",
              "P0808",
              "P0809",
              "P0902"
            ],
            "example": "P0802"
          },
          "notes": {
            "type": "string",
            "description": "Free-text notes",
            "example": "Partial payment against INV-2026-0007"
          }
        },
        "required": [
          "invoiceId",
          "amountReceived",
          "currency",
          "receivedDate",
          "purposeCode"
        ]
      }
    }
  }
}